do i look stressed to you
the post.html and/or page.html isnt picking up on any of my css sheets. i tried to change them int scss and throw them into the _sass folder but that didnt work. i even named them under the minima names to convince the Computer In The Sky that they're the same shit as before and should be used, but that didnt work. im frustrated. also i tried to "commit to master" more often but after tinkering for a while, i realized i was committing to master in the other repo. fuck me. rtfm, wiks. teehee